Split pending txs storage on interface and implementation
